book快捷鍵功能完成
